Explain ROUGE-L instead of ROUGE-SU4
I would really like to use ROUGE-SU4 as this is also what some
of the other papers on summarization are using.

However, the library (that I use) to calculate the ROUGE scores
does not support it and the original PERL script is pretty
complicated to setup. As I don't have much time left, I will
just keep using the working library and not try to use a different
one just to get the SU4 scores...